requirements.txt
openai==1.43.0
anthropic==0.34.2
python-dotenv==1.0.1
rich==13.7.1
sounddevice==0.4.6
elevenlabs==1.8.1
supabase==2.8.1
pinecone-client==3.0.3
langchain==0.3.0
cohere==5.11.0
PyYAML==6.0.2
httpx==0.27.2
google-cloud-storage==2.18.2
google-cloud-pubsub==2.31.1